+++ packages/emitter-astro/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @contentrain/emitter-astro
 
+## 0.18.2
+
+### Patch Changes
+
+- Updated dependencies [53505cd]
+  - @contentrain/types@1.24.0
+
 ## 0.18.1
 
 ### Patch Changes